+++ components/index.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Amplify from 'aws-amplify';
-import App from './App';
+import AlgosByCategories from './AlgosByCategories';
 import config from '../src/aws-exports'
 Amplify.configure(config)
 
 ReactDOM.render(
   <div>
-    <h1>Hello, ALGO!!</h1>
-    <App />
+    < AlgosByCategories />
   </div>,
   document.getElementById('root')
 );
